--- src/main/java/AddCommand.java
@@ -1,2 +1,0 @@
-public class AddCommand {
-}
