+++ src/main/java/goblin/command/MarkCommand.java
@@ -1,7 +1,15 @@
-public class UnmarkCommand extends Command{
+package goblin.command;
+
+import goblin.Storage;
+import goblin.Ui;
+import goblin.task.Task;
+import goblin.TaskList;
+import goblin.exception.OrkException;
+
+public class MarkCommand extends Command {
     protected int index;
 
-    public UnmarkCommand(int index) {
+    public MarkCommand(int index) {
         super();
         this.index = index;
     }
@@ -12,8 +20,8 @@ public class UnmarkCommand extends Command{
             if (index + 1 > TaskList.list.size()) {
                 throw new OrkException("No such task");
             }
-            task.undone();
-            Ui.printDoneMessage(task);
+            task.done();
+            ui.printDoneMessage(task);
             Storage.writeToDisk(tasks);
         } catch (OrkException exception) {
             ui.printException(exception);
