+++ src/main/java/goblin/task/Task.java
@@ -1,3 +1,5 @@
+package goblin.task;
+
 public class Task {
     protected String description;
     protected boolean isDone;
