+++ src/main/java/goblin/task/ToDos.java
@@ -1,4 +1,9 @@
-public class ToDos extends Task{
+package goblin.task;
+
+
+import goblin.task.Task;
+
+public class ToDos extends Task {
     public ToDos(String description) {
         super(description);
     }
